+++ src/components/checkout/CouponInsertionCard.tsx
@@ -3,11 +3,12 @@ import { CardProps } from 'antd/lib/card'
 import { FormComponentProps } from 'antd/lib/form'
 import axios from 'axios'
 import { useAuth } from 'lodestar-app-element/src/contexts/AuthContext'
-import React, { useState } from 'react'
+import React, { useContext, useEffect, useState } from 'react'
 import { useIntl } from 'react-intl'
 import styled from 'styled-components'
 import { handleError } from '../../helpers'
 import { checkoutMessages, codeMessages, commonMessages } from '../../helpers/translation'
+import { AuthModalContext } from '../auth/AuthModal'
 import AdminCard from '../common/AdminCard'
 import { BREAK_POINT } from '../common/Responsive'
 import messages from './translation'
